Adds manual feed contract that adheres to chainlink interface and uses exchanges rates from feed. running prettier on sol files. new take on deployment to deploy mulitple instances of the same contract

+++ contracts/token/ISPEND.sol
@@ -1,28 +1,27 @@
 pragma solidity 0.5.17;
 
-
 interface ISPEND {
-     /**
-     * @dev Returns the amount of tokens in existence.
-     */
-    function totalSupply() external view returns (uint256);
+  /**
+   * @dev Returns the amount of tokens in existence.
+   */
+  function totalSupply() external view returns (uint256);
 
-    /**
-     * @dev Returns the amount of tokens owned by `account`.
-     */
-    function balanceOf(address account) external view returns (uint256);
+  /**
+   * @dev Returns the amount of tokens owned by `account`.
+   */
+  function balanceOf(address account) external view returns (uint256);
 
-    /**
-     * @dev mint token for account, only call by minters
-     */
-    function mint(address account, uint amount) external returns(bool);
+  /**
+   * @dev mint token for account, only call by minters
+   */
+  function mint(address account, uint256 amount) external returns (bool);
 
-    /**
-     * @dev burn token for account, only call by minters
-     */
-    function burn(address account, uint amount) external returns(bool);
+  /**
+   * @dev burn token for account, only call by minters
+   */
+  function burn(address account, uint256 amount) external returns (bool);
 
-    event Mint(address account, uint amount);
+  event Mint(address account, uint256 amount);
 
-    event Burn(address account, uint amount);
+  event Burn(address account, uint256 amount);
 }
